refactored task retrieval location
